feat: add RocketQADualEncoder
