+++ examples/jina_example/flows/query.yml
@@ -6,13 +6,13 @@ with:
   cors: True
 executors:
   - name: du_encoder
-    uses: DualEncoder
+    uses: RocketQADualEncoder
     py_module:
-        - executors.py
+        - rocketqa_encoder/executor.py
     uses_with:
         model: v1_marco_de
         use_cuda: $JINA_USE_CUDA
-        device_id: 1
+        device_id: 0
         batch_size: 1
     timeout_ready: '-1'
   - name: vec_indexer
@@ -26,6 +26,6 @@ executors:
     uses_with:
         model: v1_marco_de
         use_cuda: $JINA_USE_CUDA
-        device_id: 0
+        device_id: 1
         batch_size: 32
     timeout_ready: '-1'
