+++ examples/jina_example/rocketqa_encoder/.gitignore
@@ -1,0 +1,1 @@
+token
