+++ examples/jina_example/rocketqa_encoder/README.md
@@ -1,0 +1,4 @@
+# RocketQADualEncoder
+
+This is a wrapper a RocketQA library provided by [Baidu NLP Team](https://github.com/PaddlePaddle/RocketQA).
+
