+++ examples/jina_example/rocketqa_encoder/config.yml
@@ -1,0 +1,4 @@
+jtype: RocketQADualEncoder
+metas:
+  py_modules:
+    - executor.py
