+++ examples/jina_example/rocketqa_encoder/manifest.yml
@@ -1,0 +1,3 @@
+manifest_version: 1
+name: RocketQADualEncoder
+keywords: qa, text, PaddlePaddle, en, zh, encoder
