+++ examples/jina_example/rocketqa_encoder/requirements.txt
@@ -1,0 +1,2 @@
+git+https://github.com/sfwydyc/rocketqa.git
+paddlepaddle>=2.2.0
